+++ amethyst/src/fdroid/java/com/vitorpamplona/amethyst/service/notifications/PushNotificationUtils.kt
@@ -21,7 +21,7 @@
 package com.vitorpamplona.amethyst.service.notifications
 
 import com.vitorpamplona.amethyst.AccountInfo
-import com.vitorpamplona.amethyst.service.retryIfException
+import com.vitorpamplona.amethyst.commons.util.retryIfException
 import kotlinx.coroutines.Dispatchers
 import okhttp3.OkHttpClient
